+++ utils/dynamic_buffer.h
@@ -1,6 +1,6 @@
 /**
  * @file dynamic_buffer.h
- * @author Michal Cyprian
+ * @author Michal Cyprian <xcypri01@stud.fit.vutbr.cz>
  *
  * @section DESCRIPTION
  *
@@ -13,10 +13,10 @@ typedef struct {
     char *buffer;
 } TDynamic_buffer;
 
-int init_buffer(TDynamic_buffer *b, unsigned initial_length);
-int realloc_buffer(TDynamic_buffer *b, unsigned size);
+int init_buffer(TDynamic_buffer *b, size_t initial_length);
+int realloc_buffer(TDynamic_buffer *b, size_t size);
 void free_buffer(TDynamic_buffer *b);
 int add_char(TDynamic_buffer *b, char c);
 int add_str(TDynamic_buffer *b, char* word);
 char *read_buffer(TDynamic_buffer *b);
-int empty_buffer(TDynamic_buffer *b, unsigned initial_length);
+int empty_buffer(TDynamic_buffer *b, size_t initial_length);
